fix: split marshal tests
